requirement.txt
ccxt
pandas
pandas_ta
pprint36
moment
python-dotenv
